requirements.txt
gspread
py-trello
pytz
python-telegram-bot